refactor: console validation hook 파일 분리
