+++ src/console/useConsoleValidation.js
@@ -1,0 +1,12 @@
+import useValidation from "../validation/useValidation.js";
+
+const executeValidations = (input, rules) => {
+  for (const rule of rules) {
+    rule(input);
+  }
+  return input;
+};
+
+const useConsoleValidation = useValidation(executeValidations);
+
+export default useConsoleValidation;
